Try FIX openhems-addon by specifying the  typing-extensions version? (#31)
Upgrade to 0.3.2
Try FIX openhems-addon by specifying the typing-extensions version?
Upgrade sub-actions versions

+++ test_venv.sh
@@ -1,8 +1,28 @@
 #!/bin/bash
 
-rm -rf dist
-uv build
-uv venv --clear test_env
-source test_env/bin/activate
-uv pip install dist/openhems-*.whl
+if [ $# -ne 1 ]; then
+  echo "Missing argument, expecting 'action' : 'local' or 'remote'"
+  exit 1
+fi
+
+action=$1
+
+if [ "$action" = "remote" ]; then
+    echo "Run OpenHEMS server with the latest version from PyPI"
+    uv venv --clear test_env
+    source test_env/bin/activate
+    uv pip install openhems
+    uv pip show openhems
+else if [ "$action" = "local" ]; then
+    echo "Run OpenHEMS server with the local version"
+    rm -rf dist
+    uv build
+    uv venv --clear test_env
+    source test_env/bin/activate
+    uv pip install dist/openhems-*.whl
+else
+  echo "Wrong argument ($action) expecting 'local' or 'remote'"
+  exit 1
+fi; fi
+
 ./test_venv.py
